Listen for work clearing message
This should allow us to handle the case where work never gets assigned
due to an error on beak's end (beak will send a message to clear
work). The more robust fix will be for schrute to periodically check
worker status with Rainforest, but that's a bit trickier and will be
slower in any case.

Fixes #508