Hard-code polling endpoint
Setting it dynamically made a lot of sense when it was hard to deploy
the extension but not so much any more. Polling will be nuked soon in
any case, so it doesn't make sense to have to deal with error-proneness
for the time being.

See #629 (which I'm not 100% sure is fully fixed).