Ensure load order in state restoration
The local storage (which stores worker state) was coming back before the
sync storage (which stores authentication) in some circumstances. That
led to trying to set the worker state to "ready" before authenticating,
which doesn't work (it puts the worker back to "inactive"). Promisifying
the storage handlers should fix the issue.