Retry "ready" state updates
When the extension automatically reconnects after a connection failure, it sends
its current state to the server. However, it seems that under certain
circumstances the server will keep around a "zombie" connection (that could be
"ready") until it times out, so the new connection can't move to "ready" until
the old connection clears (since we enforce that only one connection can be
"ready" per worker). In this case, it makes sense to retry for a while instead
of alerting the user. (Manual state updates won't retry under any
circumstances.)

My only concern is that this introduces a mismatch between the extension state
and the server state, which could be hard to debug if something goes wrong. But
I can't think of a better alternative.

Fixes #752